Add the unittest infrastructure
Firefox addons has no unittest framework.

As the Firefox addon system has no a unittest framework, it is
impossible to load scripts of the addon, because these scripts have
references to the internal browser structures.

So, the strategy for testing anything is to put functions and classes
to a separate js-file and connect the file to the addon in the
contents scripts section.

After that, functions and classes become visible in the addon and may
be used by different addon parts.

Meanwhile, the unittest system (based on mocha and etc.) can get
access to this separate file with functions and classes and test them.