Change the method for replacing a videoframe
In different versions of Firefox videoframes are replaced
differently. So in new version of Firefox 75.0.2 the replacement
option is undefined for a while, thus the videoframe left initial.

Now the replacement code doesn't guess time when the replacement
option is ready. Now it replaces the videoframe periodically, so the
option doesn't need to be set in exact time.